Remove the interface MinimalProjectModel and rename the DefaultMinimalProjectModel to MinimalProjectModel.
* Splitting into implementation and interface is not needed here. (The MinimalProjectModel is for now only meant for testing not part of a plugin API.)
